refactor 2

+++ services/handbook/src/CC.HandbookService/Infrastructure/Services/HandbookLoader.cs
@@ -2,53 +2,29 @@ using System.Globalization;
 using CC.HandbookService.Application.Dependencies;
 using CC.HandbookService.Domain.Handbooks;
 using CC.HandbookService.Infrastructure.Persistence;
+using CC.HandbookService.Infrastructure.Services.CsvParsing;
 using CC.HandbookService.Infrastructure.Services.CsvParsing.HandbookMaps;
 using CsvHelper;
 using FluentResults;
 using Microsoft.EntityFrameworkCore;
 using MissingFieldException = CsvHelper.MissingFieldException;
 
-namespace CC.HandbookService.Infrastructure.Services.CsvParsing;
+namespace CC.HandbookService.Infrastructure.Services;
 
-public class HandbookLoader(DatabaseContext context) : IHandbookLoader
+public class HandbookLoader<T> (
+    IHandbookParser parser,
+    DatabaseContext context) : IHandbookLoader 
+    where T : HandbookItem 
 {
-    public async Task<Result> LoadAsync<T>(Stream stream) where T : HandbookItem
+    public async Task<Result> LoadAsync(Stream stream)
     {
-        var streamReader = new StreamReader(stream);
-        using var csvReader = new CsvReader(streamReader, CultureInfo.InvariantCulture);
-        csvReader.Context.RegisterClassMap<HandbookItemMap>();
+        var result = await parser.ParseAsync<T>(stream);
+        if (!result.IsSuccess)
+            return Result.Fail(result.Errors);
         
-        var records = new List<T>();
-        try
-        {
-            await foreach (var record in csvReader.GetRecordsAsync<T>())
-                records.Add(record);
-        }
-        catch (MissingFieldException ex)
-        {
-            var cellName = TryGetCellName(ex);
-            return cellName != null 
-                ? Result.Fail($"Пропущенно поле в ячейке {cellName}") 
-                : Result.Fail("Непредвиденная ошибка");
-        }
-        catch (ValidationException ex)
-        {
-            var cellName = TryGetCellName(ex);
-            return cellName != null
-                ? Result.Fail($"Не заполнено обязательное поле в ячейке {cellName}")
-                : Result.Fail("Непредвиденная ошибка");
-        }
-            
         await context.Set<T>().ExecuteDeleteAsync();
-        context.Set<T>().AddRange(records); 
+        context.Set<T>().AddRange(result.Value); 
 
         return Result.Ok();
-    }
-    
-    private static string? TryGetCellName(CsvHelperException ex)
-    {
-        if (ex.Context == null || ex.Context.Parser == null || ex.Context.Reader == null)
-            return null;
-        return CsvCellHelper.GetCellName(ex.Context.Parser.Row, ex.Context.Reader.CurrentIndex + 1);
     }
 }
